backend/requirements.txt
# Django Core
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Authentication
djangorestframework-simplejwt==5.3.0
PyJWT==2.8.0

# Async Tasks
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0

# Image Processing
Pillow==10.1.0

# Environment Variables
python-decouple==3.8

# Testing
pytest==7.4.3
pytest-django==4.7.0
hypothesis==6.92.1
factory-boy==3.3.0

# Code Quality
black==23.11.0
flake8==6.1.0
isort==5.12.0

# API Documentation
drf-spectacular==0.26.5

# Filtering
django-filter==23.5
djangorestframework-gis==1.0

# Geospatial
django-geojson==4.0.0

# Utilities
python-dateutil==2.8.2

# Push Notifications (optional)
# firebase-admin==6.3.0
